requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
twilio==6.45.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
